+++ CHANGES.md
@@ -16,6 +16,7 @@
 - Outline: Significantly update `cratedb-outline.yaml`
 - Bundle: Started accepting `--url`/`ABOUT_OUTLINE_URL` option to specify
   alternative input outline file
+- Bundle: Improved handling of `--format` option
 
 ## v0.0.3 - 2025-05-10
 - Outline: Refactored the source of truth for the documentation outline